modelagem-preditiva-main/api_/api/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
